@@ -4,8 +4,8 @@ description: WidgetFactory extension to render IFRAME with the official WebView 
 homepage: https://github.com/daohoangson/flutter_widget_from_html
 
 environment:
-  flutter: ">=3.0.0"
-  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.7.0"
+  sdk: ">=2.19.0 <4.0.0"
 
 dependencies:
   flutter:
