@@ -218,16 +218,13 @@ class StyleSizing {
 
 class _MinWidthZero extends ConstraintsTransformBox {
   const _MinWidthZero({
-    required Widget child,
-    Key? key,
+    super.child,
     required TextDirection textDirection,
   }) : super(
           alignment: textDirection == TextDirection.ltr
               ? Alignment.topLeft
               : Alignment.topRight,
           constraintsTransform: transform,
-          key: key,
-          child: child,
         );
 
   static BoxConstraints transform(BoxConstraints bc) =>
