@@ -82,6 +82,5 @@ class StyleTextAlign {
 }
 
 class _TextAlignCenter extends Center {
-  const _TextAlignCenter(Widget child, {Key? key})
-      : super(child: child, heightFactor: 1.0, key: key);
+  const _TextAlignCenter(Widget child) : super(child: child, heightFactor: 1.0);
 }
