@@ -55,8 +55,7 @@ class TagA {
 class _TagABit extends BuildBit<GestureRecognizer?, GestureRecognizer> {
   final GestureTapCallback onTap;
 
-  const _TagABit(BuildTree? parent, TextStyleBuilder tsb, this.onTap)
-      : super(parent, tsb);
+  const _TagABit(super.parent, super.tsb, this.onTap);
 
   @override
   bool? get swallowWhitespace => null;
