@@ -391,7 +391,7 @@ class _ColumnTestApp extends StatelessWidget {
   final String? html;
   final Key? keyBottom;
 
-  const _ColumnTestApp({this.html, Key? key, this.keyBottom}) : super(key: key);
+  const _ColumnTestApp({this.html, this.keyBottom});
 
   @override
   Widget build(BuildContext _) => Scaffold(
@@ -413,7 +413,7 @@ class _ColumnTestApp extends StatelessWidget {
 class _ListViewTestApp extends StatelessWidget {
   final String? html;
 
-  const _ListViewTestApp({this.html, Key? key}) : super(key: key);
+  const _ListViewTestApp({this.html});
 
   @override
   Widget build(BuildContext _) => Scaffold(
@@ -429,7 +429,7 @@ class _ListViewTestApp extends StatelessWidget {
 class _SliverListTestApp extends StatelessWidget {
   final Key? keyBottom;
 
-  const _SliverListTestApp({Key? key, this.keyBottom}) : super(key: key);
+  const _SliverListTestApp({this.keyBottom});
 
   @override
   Widget build(BuildContext _) => Scaffold(
