@@ -486,8 +486,7 @@ class _BuildBitWidgetFactory extends WidgetFactory {
 }
 
 class _InputBuildContextBit extends BuildBit<BuildContext, Widget> {
-  const _InputBuildContextBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _InputBuildContextBit(super.parent, super.tsb);
 
   @override
   bool get isInline => false;
@@ -501,8 +500,7 @@ class _InputBuildContextBit extends BuildBit<BuildContext, Widget> {
 }
 
 class _InputGestureRecognizerBit extends BuildBit<GestureRecognizer?, dynamic> {
-  const _InputGestureRecognizerBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _InputGestureRecognizerBit(super.parent, super.tsb);
 
   @override
   dynamic buildBit(GestureRecognizer? recognizer) {
@@ -520,8 +518,7 @@ class _InputGestureRecognizerBit extends BuildBit<GestureRecognizer?, dynamic> {
 }
 
 class _InputVoidBit extends BuildBit<void, dynamic> {
-  const _InputVoidBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _InputVoidBit(super.parent, super.tsb);
 
   @override
   dynamic buildBit(void _) => '(null)';
@@ -532,8 +529,7 @@ class _InputVoidBit extends BuildBit<void, dynamic> {
 }
 
 class _InputTextStyleHtmlBit extends BuildBit<TextStyleHtml, InlineSpan> {
-  const _InputTextStyleHtmlBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _InputTextStyleHtmlBit(super.parent, super.tsb);
 
   @override
   InlineSpan buildBit(TextStyleHtml tsh) =>
@@ -545,8 +541,7 @@ class _InputTextStyleHtmlBit extends BuildBit<TextStyleHtml, InlineSpan> {
 }
 
 class _OutputGestureRecognizerBit extends BuildBit<void, GestureRecognizer> {
-  const _OutputGestureRecognizerBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _OutputGestureRecognizerBit(super.parent, super.tsb);
 
   @override
   GestureRecognizer buildBit(void _) => MultiTapGestureRecognizer();
@@ -557,8 +552,7 @@ class _OutputGestureRecognizerBit extends BuildBit<void, GestureRecognizer> {
 }
 
 class _OutputInlineSpanBit extends BuildBit<void, InlineSpan> {
-  const _OutputInlineSpanBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _OutputInlineSpanBit(super.parent, super.tsb);
 
   @override
   InlineSpan buildBit(void _) => const WidgetSpan(child: Text('foo'));
@@ -569,8 +563,7 @@ class _OutputInlineSpanBit extends BuildBit<void, InlineSpan> {
 }
 
 class _OutputStringBit extends BuildBit<void, String> {
-  const _OutputStringBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _OutputStringBit(super.parent, super.tsb);
 
   @override
   String buildBit(void _) => 'foo';
@@ -581,8 +574,7 @@ class _OutputStringBit extends BuildBit<void, String> {
 }
 
 class _OutputWidgetBit extends BuildBit<void, Widget> {
-  const _OutputWidgetBit(BuildTree? parent, TextStyleBuilder tsb)
-      : super(parent, tsb);
+  const _OutputWidgetBit(super.parent, super.tsb);
 
   @override
   bool get isInline => false;
