@@ -33,11 +33,11 @@ class AudioPlayer extends StatefulWidget {
   const AudioPlayer(
     this.url, {
     this.autoplay = false,
-    Key? key,
+    super.key,
     this.loop = false,
     this.muted = false,
     this.preload = false,
-  }) : super(key: key);
+  });
 
   @override
   State<AudioPlayer> createState() => _AudioPlayerState();
@@ -140,12 +140,11 @@ class _PlayButton extends StatelessWidget {
   final Stream<bool> stream;
 
   const _PlayButton({
-    Key? key,
     required this.pause,
     required this.play,
     required this.size,
     required this.stream,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext _) => StreamBuilder<bool>(
@@ -168,12 +167,11 @@ class _PositionText extends StatelessWidget {
   final double size;
 
   const _PositionText({
-    Key? key,
     required this.durationStream,
     required this.isNarrow,
     required this.positionStream,
     required this.size,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext _) => StreamBuilder<Duration?>(
@@ -215,12 +213,11 @@ class _PositionSlider extends StatelessWidget {
   final double size;
 
   const _PositionSlider({
-    Key? key,
     required this.durationStream,
     required this.positionStream,
     required this.seek,
     required this.size,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) => StreamBuilder<Duration?>(
@@ -258,11 +255,10 @@ class _MuteButton extends StatelessWidget {
   final Stream<double> stream;
 
   const _MuteButton({
-    Key? key,
     required this.setVolume,
     required this.size,
     required this.stream,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext _) => StreamBuilder<double>(
